Create #ReviseWithArsh
https://leetcode.com/problems/bulls-and-cows/submissions/870331892/
Leetcode_Microsoft_/#ReviseWithArsh

# Intuition
<!-- Describing  first thoughts on how to solve this problem. -->
In the question they are basically talking about a game Bulls and Cows in which suppose you and your friend are playing a game in which you have written down a secret number in a paper and now you are asking your friend to make a guess...
Suppose you have written a secret number "4566" and now you asked your friend to guess. So there are two possibilities:
1.Either your friend will guess BULL.
2.Or he/she will guess COW.

THINGS TO UNDERSTAND WHAT IS BULL AND COW MEANS:
BULL denotes the digits of guess number which are also present in secret number and are at the right position as the digit in the secret number.
COWS denotes the digits of guess number which are also in the secret number but not at the right position as in the secret number.

# Approach
<!-- Describing the approach to solving the problem. -->
First of all, we will consider two arrays:
1. An array of SecretFrequency of integer type which will contain occurrences of a particular number in the secret number.
2. Another array of GuessFrequecy of integer type which will contain occurrences of the digits in the guess number.

Firsty we initilaize Both COWS and BULLS as 0 then call the loop and check if both digit in the guess number is a COW or BULL if both position and value is correct, we will increment the value of BULLS
IF NOT
THEN, suppose if the digit on the first index in the secret and guess string does not match then we will increment both the arrays 
(i). Array SecretFre with the digit on the (suppose)first index of secret string.
(ii).Array GuessFre with the digit on the (suppose)first index of guess string.

THEN we will find the number of COWS by finding the mininmum occurence of the number in both the arrays.

THEN FINALLY WILL RETURN THE ANS IN THE FORMAT "xAyB"
where x is the number of BULLS and y is the number of COWS...

REMEMBER WE HAVE DIGITS[0,9] i.e. 10 digits
ALSO LINE 16 AND 17 , it is done to convert string into integer because we have to store integer in the aaray...