Create #ReviseWithArsh
LEETCODE_MICROSOFT_7#ReviseWithArsh
https://leetcode.com/problems/course-schedule/submissions/871823121/
Intuition
In this first of all we need to understand what we are given here,
No. of courses
Prerequisites i.e. say[1,2] so in this you need to complete two tasks and inn order to complete task 1 you have to complete task 2...
NOW what we need to do is to see whether can finish both courses or not depending on the prerequisites.
If can finish then return true otherwise false...

Approach
In this in order to check whether can finish courses or not, we will make the adjacency list with every element choosing one at a time and keep on adding the elements dependent correspondingly to the chosen element
Then will apply DFS on the given set and trace the elements not visited one at a time and on the element with no longer attachments, we will mark this current element as complete and trace the previous path again in order to visit the remaining connected elements
After cosidering this current element as complete will go to the previous element and check if it is connected with any other element if yes then check if that connected element is complete or not if complete the forward edge is there otherwise back edge...