Create #ReviseWithArsh
LEETCODE/MICROSOFT_11/#ReviseWithArsh
https://leetcode.com/problems/number-of-ways-to-arrive-at-destination/submissions/874456784/
Intuition
Here we are given n intersections i.e. no. of nodes starting from 0 to n-1 and there are bidirectional roads between the intersections(nodes).
Also, we are given a 2D integer array roads which is having three values; ui, vi, timei where ui and vi nodes denoting that there is a path between the two; timei is the time taken to travel from node ui to vi or vi to ui.
WE NEED TO FIND THE NUMBER OF WAYS THAT SO THAT THAT WE CAN REACH TO OUR DESTINATION IN SHORTEST AMOUNT OF TIME

Approach
In order to count the number of ways to reach to the destination in smallest amount of amount of time we will take an integer vector which has 3 values u and v denoting the next node and w is the weighted edge
Then we have taken the vector time and vector way and also a priority queue in which we push the pair of timeTotravel and adjacent node starting from{0,0} will pop pot until the pq becomes empty.

Once we pop out one pair then will go to the adjaceny node and will check if the current time+ time to go to adjacent node <=time at the adjacent node then we will update it and then will push the pair in the queue and if == time at the adjacent node then we will add the no.of ways to reach the same node...