Create #ReviseWithArsh
LEETCODE/MICROSOFT_12/#ReviseWithArsh
https://leetcode.com/problems/longest-happy-prefix/submissions/875805375/

Intuition
In this question we are goven a string from which we need to find the happy string i.e. a string which can be wirk as both prefix and suffix
Our task is to find the longest happy string...

Approach
( LPS array algorithm)

LPS is preprocessing part of KPM alogorithm. In this approach basically we calculate values in lps[]. To do that, we keep track of the length of the longest prefix suffix value (we use j variable for this purpose) for the previous index. We initialize lps[0] and j as 0. If s[j] and s[i] match, we increment j by 1 and assign the incremented value to lps[i]. If s[i] and s[j] do not match and j is not 0, we update j to lps[j-1].

Complexity
Time complexity: O(N)
Space complexity: O(N)